xfpga: rename opae_ioctl files to opae_drv (#1079)
* xfpga: rename opae_ioctl files to opae_drv

opae_drv will be used as the driver interface for both sysfs and ioctl
