+++ libopae/plugins/xfpga/CMakeLists.txt
@@ -53,7 +53,7 @@ set(SRC
   event.c
   properties.c
   init.c
-  opae_ioctl.c
+  opae_drv.c
   sysfs.c
   wsid_list.c
   token_list.c
