+++ libopae/plugins/xfpga/opae_drv.c
@@ -35,7 +35,7 @@
 #include <opae/fpga.h>
 #include <safe_string/safe_string.h>
 #include "common_int.h"
-#include "opae_ioctl.h"
+#include "opae_drv.h"
 #include "intel-fpga.h"
 #include "fpga-dfl.h"
