+++ libopae/plugins/xfpga/umsg.c
@@ -28,7 +28,7 @@
 #include "opae/utils.h"
 #include "opae/umsg.h"
 #include "common_int.h"
-#include "opae_ioctl.h"
+#include "opae_drv.h"
 #include "intel-fpga.h"
 
 #include <sys/types.h>
